# server/src/docs/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: StrmScan APIs
  description: StrmScan APIs
  license:
    name: Github
    url: 'https://github.com/strmchain/strmscan'
...
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/rewards/{voter}:
    get:
      tags:
        - Rewards
      summary: Get list rewards of a voter
      parameters:
        - name: voter
          in: path
          required: true
          description: account address
          type: number
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
  /api/rewards/epoch/{epochNumber}:
    get:
      tags:
        - Rewards
      summary: Get list reward of an epoch
      parameters:
        - name: epochNumber
          in: path
          required: true
          description: epoch number
          type: number
        - name: hash
          in: path
          required: true
          description: account address
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
  /api/rewards/total/{address}/{fromEpoch}/{toEpoch}:
    get:
      tags:
        - Rewards
      summary: Get total rewards of an address
      parameters:
        - name: address
          in: path
          required: true
          description: account address
          type: string
        - name: fromEpoch
          in: path
          required: true
          description: from epoch
          type: number
        - name: toEpoch
          in: path
          required: true
          description: to epoch
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
...